+++ components/media/Media.vue
@@ -1,5 +1,5 @@
 <template>
-  <component :is="componentsPrefix + componentName" :media-category="mediaCategory" :media-uri-prop="mediaUriProp" :authors-uri-prop="authorsUriProp" :project-id-prop="projectIdProp" :auth-token="authToken" :sections-user-id-prop="sectionsUserIdProp" :media-translation-prefix="mediaTranslationPrefix" :show-create-media-button="showCreateMediaButton" :media-response-prop="mediaResponseProp" :media-by-id-uri-prop="mediaByIdUriProp" :media-by-id-response-prop="mediaByIdResponseProp" :media-id-prop="mediaId" :create-media-path="createMediaPath" :edit-media-path="editMediaPath" :medias-path="mediasPath" :bo-usage="boUsage" :access-limited="accessLimited" @updateMediaComponent="onMediaComponentUpdate" :with-select-media-button="withSelectMediaButton" :nuxt-sections="nuxtSections" :is-create-media="isCreateMedia" :media-id-editing="mediaIdEditing" @onMediaSelected="(media) => $emit('getSelectedMedia', media)"  />
+  <component :is="componentsPrefix + componentName" :media-category="mediaCategory" :media-uri-prop="mediaUriProp" :authors-uri-prop="authorsUriProp" :project-id-prop="projectIdProp" :auth-token="authToken" :sections-user-id-prop="sectionsUserIdProp" :media-translation-prefix="mediaTranslationPrefix" :show-create-media-button="showCreateMediaButton" :media-response-prop="mediaResponseProp" :media-by-id-uri-prop="mediaByIdUriProp" :media-by-id-response-prop="mediaByIdResponseProp" :media-id-prop="mediaId" :create-media-path="createMediaPath" :edit-media-path="editMediaPath" :medias-path="mediasPath" :bo-usage="boUsage" :access-limited="accessLimited" @updateMediaComponent="onMediaComponentUpdate" :with-select-media-button="withSelectMediaButton" :nuxt-sections="nuxtSections" :is-create-media="isCreateMedia" :media-id-editing="mediaIdEditing" :applied-filters="appliedFilters" :folder-type="folderType" @onMediaSelected="(media) => $emit('getSelectedMedia', media)"  />
 </template>
 
 <script>
@@ -10,7 +10,9 @@ export default {
     return {
       componentName: 'MediaListMedias',
       mediaId: '',
-      isCreateMedia: false
+      isCreateMedia: false,
+      appliedFilters: '',
+      folderType: ''
     }
   },
   props: {
@@ -148,6 +150,8 @@ export default {
       this.componentName = component.name
       this.mediaId = component.mediaId
       this.isCreateMedia = component.isCreateMedia
+      this.appliedFilters = component.appliedFilters
+      this.folderType = component.folderType
     }
   }
 }