+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 ---
 
 
+### v1.0.5 (16/11/2023)
+
+---
+
+- [x] Add Mobile responsiveness support for the components
+
+- [x] Fix for Media component filter to preserve filters
+
+
 ### v1.0.4 (24/8/2023)
 
 ---